E3_DASHBOARD-5/requirements.txt
streamlit
pandas
seaborn
matplotlib
scikit-learn
scipy
numpy
datetime
PIL
plotly
itertools
statsmodels.api
